+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Search Instances method
 - List Instances method
 - List Access Tokens method
+- Search Data Tables method
 
 ### Removed
 - Find Users method (Breaking change)